Major Update 16 January 2018
-EMG_plot and PAS_bar now takes a clean combination of input arguments
and has a varargin to specify which session user would like to plot

-New function fix_snip_epoch_mismatch.m was written to manually excise
snips and epocs if they are mismatched at data collection.

-MASTER_periop.m may need to be deprecated now, but essentially it just
runs the conversion function to get the MasterDataset containing all the
data. Will likely need to clean house and rename it soon

-Modified maximizer.m to play nice with the other functions

-Changed initparams.m to the initial value of EMG = [1:2]

-Modified TDT_preproc.m to check for snip/epoc mismatch and to trigger
exit errors if these are the case, permitting human intervention